Revert "LPS-85787 Merge FilterParser into FilterProvider for the same reasons."
This reverts commit 7894b3f35d18b024984adb10d1688ff428e2a961.
